tests/wpt/web-platform-tests/idle-detection/interceptor.https.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/inexorabletash/idle-detection">
<title>Tests the Idle Detection API</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="/gen/mojo/public/mojom/base/string16.mojom.js"></script>
<script src="/gen/mojo/public/mojom/base/time.mojom.js"></script>
<script src="/gen/third_party/blink/public/platform/modules/idle/idle_manager.mojom.js"></script>
<script src="mock.js"></script>
<script>
'use strict';

promise_test(async t => {
  // Basic test that expects navigator.idle.query() to call internally
  // addMonitor, which in turn will return an ACTIVE state.
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      return Promise.resolve({
        state: {
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.LOCKED
        }
      });
  });

  let status = await navigator.idle.query({threshold: 10});

  assert_equals(status.state.user, "active");
  assert_equals(status.state.screen, "locked");
}, 'query()');

promise_test(async t => {
  // Verifies that an event is thrown when a change of state from IDLE to ACTIVE
  // is detected.
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      t.step_timeout(() => {
        monitorPtr.update({
          user: UserIdleState.IDLE,
          screen: ScreenIdleState.UNLOCKED
        });
      }, 0);
      return Promise.resolve({
        state: {
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.UNLOCKED
        }
      });
    });

  let monitor = await navigator.idle.query({threshold: 10});

  await new EventWatcher(t, monitor, ["change"]).wait_for("change");

  assert_equals(monitor.state.user, "idle");
  assert_equals(monitor.state.screen, "unlocked");
}, 'updates once');


promise_test(async t => {
  // Simulates the user being active, going idle and then going back active
  // again.
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      // Updates the client once with the user idle.
      t.step_timeout(() => {
        monitorPtr.update({
          user: UserIdleState.IDLE,
          screen: ScreenIdleState.UNLOCKED
        });
      }, 0);
      // Updates the client a second time with the user active.
      t.step_timeout(() => {
        monitorPtr.update({
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.UNLOCKED
        });
      }, 1);
      return Promise.resolve({
        state: {
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.UNLOCKED
        }
      });
    });

  let monitor = await navigator.idle.query({threshold: 10});

  let watcher = new EventWatcher(t, monitor, ["change"]);

  // waits for the first event.
  await watcher.wait_for("change");
  assert_equals(monitor.state.user, "idle");

  // waits for the second event.
  await watcher.wait_for("change");
  assert_equals(monitor.state.user, "active");
}, 'updates twice');

promise_test(async t => {
  // Simulates a locked screen.
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      return Promise.resolve({
        state: {
          user: UserIdleState.ACTIVE,
          screen: ScreenIdleState.LOCKED
        }
      });
    });

  let monitor = await navigator.idle.query({threshold: 10});

  assert_equals(monitor.state.screen, "locked");
}, 'locked screen');

promise_test(async t => {
  // Simulates the service becoming unavailable.
  expect(addMonitor).andReturn((threshold, monitorPtr) => {
      return new Promise((resolve, reject) => {
        // leave the renderer deliberately hanging by not resolve()-ing.
      });
    });

  let error = new Promise((resolve, reject) => {
    navigator.idle.query({threshold: 10})
      .then((e) => {reject("unexpected response :(")})
      .catch((e) => {resolve(e.message)});
  });

  // simulates what happens when the service is unavailable.
  close();

  assert_equals(await error, "Idle detection not available");
}, "service unavailable");
</script>
